Corregi funcione del buffer y lo volvi estatico. Logre que se envien los datos a ifnluxdb pero no se reciven correctamente.

+++ src/ADCData.h
@@ -4,7 +4,7 @@
 // Definición del tipo de datos ADCData
 struct ADCData {
     int pin;          // El pin del ADC
-    float value;      // El valor medido por el ADC
+    int value;      // El valor medido por el ADC
     unsigned long timestamp;  // Marca de tiempo del dato
 };
